Fix build errors
qmake-qt5 cannot be found by the configure script. This is fixed adding
the prefix of the Qt5 installation as a configure option.
Building with LLVM/Clang raises an error related to a variable sized
array. This is fixed by patching a source file to use a dynamically
allocated array instead.

Fixes #3278.